“Delete (with descendants)” action

+++ src/elements/actions/Delete.php
@@ -23,6 +23,12 @@ use craft\helpers\Db;
 class Delete extends ElementAction
 {
     /**
+     * @var bool Whether to delete the element’s descendants as well.
+     * @since 3.5.0
+     */
+    public $withDescendants = false;
+
+    /**
      * @var bool Whether to permanently delete the elements.
      * @since 3.5.0
      */
@@ -59,6 +65,10 @@ class Delete extends ElementAction
             return Craft::t('app', 'Delete permanently');
         }
 
+        if ($this->withDescendants) {
+            return Craft::t('app', 'Delete (with descendants)');
+        }
+
         return Craft::t('app', 'Delete');
     }
 
@@ -88,6 +98,12 @@ class Delete extends ElementAction
             ]);
         }
 
+        if ($this->withDescendants) {
+            return Craft::t('app', 'Are you sure you want to delete the selected {type} along with their descendants?', [
+                'type' => $elementType::pluralLowerDisplayName(),
+            ]);
+        }
+
         return Craft::t('app', 'Are you sure you want to delete the selected {type}?', [
             'type' => $elementType::pluralLowerDisplayName(),
         ]);
@@ -107,8 +123,35 @@ class Delete extends ElementAction
             }
         } else {
             $elementsService = Craft::$app->getElements();
+
+            if ($this->withDescendants) {
+                $query
+                    ->with([
+                        [
+                            'descendants',
+                            [
+                                'orderBy' => ['structureelements.lft' => SORT_DESC],
+                            ]
+                        ]
+                    ])
+                    ->orderBy(['structureelements.lft' => SORT_DESC]);
+            }
+
+            $deletedElementIds = [];
+
             foreach ($query->all() as $element) {
-                $elementsService->deleteElement($element);
+                if (!isset($deletedElementIds[$element->id])) {
+                    if ($this->withDescendants) {
+                        foreach ($element->getDescendants() as $descendant) {
+                            if (!isset($deletedElementIds[$descendant->id])) {
+                                $elementsService->deleteElement($descendant);
+                                $deletedElementIds[$descendant->id] = true;
+                            }
+                        }
+                    }
+                    $elementsService->deleteElement($element);
+                    $deletedElementIds[$element->id] = true;
+                }
             }
         }
